+++ CHANGELOG.md
@@ -20,6 +20,8 @@ All notable changes to this project will be documented in this file.
 - Added "Copy request headers".
 - Added "Copy request parameters".
 - Added "Copy response headers".
+- Added "Copy thread name".
+- Added "Copy thread group name".
 - Added thread priority to thread info of logging event.
 
 ### Changed